+++ R/eyelink_parser.R
@@ -1,6 +1,6 @@
 #' Read EyeLink ASC Files
 #'
-#' Imports data from EyeLink .ASC files into (relatively) tidy data frames for analysis and
+#' Imports data from EyeLink ASC files into (relatively) tidy data frames for analysis and
 #' visualization. Event data and/or raw sample data from the files can be imported, along with
 #' information about the tracker hardware and configuration. All data is divided into numbered
 #' blocks using the "START" and "END" messages in the ASC file.
@@ -25,7 +25,7 @@
 #' section of the EyeLink 1000 User's Guide.
 #'
 #' Note that this function cannot import EDFs directly; they must be converted to plain-text ASC
-#' using the edf2asc utilty before importing.
+#' using the edf2asc utility before importing.
 #'
 #' @usage
 #' read.asc(fname, samples = TRUE, events = TRUE, parse_all = FALSE)